+++ pom.xml
@@ -60,7 +60,7 @@
                         <!-- Default configuration for running with: mvn clean javafx:run -->
                         <id>default-cli</id>
                         <configuration>
-                            <mainClass>com.radutek.sefstudentmanager/com.radutek.sefstudentmanager.HelloApplication
+                            <mainClass>com.radutek.sefstudentmanager/loose.oose.fis.lab.student.manager.HelloApplication
                             </mainClass>
                             <launcher>app</launcher>
                             <jlinkZipName>app</jlinkZipName>
