+++ src/main/java/loose/oose/fis/lab/student/manager/HelloApplication.java
@@ -1,4 +1,4 @@
-package com.radutek.sefstudentmanager;
+package loose.oose.fis.lab.student.manager;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
