+++ src/main/java/loose/oose/fis/lab/student/manager/HelloController.java
@@ -1,4 +1,4 @@
-package com.radutek.sefstudentmanager;
+package loose.oose.fis.lab.student.manager;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
