+++ src/main/resources/loose/oose/fis/lab/student/manager/hello-view.fxml
@@ -6,7 +6,7 @@
 
 <?import javafx.scene.control.Button?>
 <VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="com.radutek.sefstudentmanager.HelloController">
+      fx:controller="loose.oose.fis.lab.student.manager.HelloController">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
     </padding>
